+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: kmatsuna <kmatsuna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/21 20:57:07 by kmatsuna          #+#    #+#             */
-/*   Updated: 2024/06/09 20:54:52 by kmatsuna         ###   ########.fr       */
+/*   Updated: 2024/06/26 18:59:00 by kmatsuna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 	dstsize = ft_strlen(dst);
 	if (size == 0 || dst == NULL)
 		return (ft_strlen(src));
-	while ((int)size - 1 > 0 && src[i] != '\0')
+	while (size - 1 > 0 && src[i] != '\0')
 	{
 		dst[i] = src[i];
 		i++;
