+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: kmatsuna <kmatsuna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 18:01:09 by kmatsuna          #+#    #+#             */
-/*   Updated: 2024/06/09 20:56:00 by kmatsuna         ###   ########.fr       */
+/*   Updated: 2024/06/26 19:07:47 by kmatsuna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
-	while (i < (int)n && (s1[i] != '\0' || s2[i] != '\0'))
+	while (i < n && (s1[i] != '\0' || s2[i] != '\0'))
 	{
 		if (s1[i] != s2[i])
 		{
